Link to the source repos
This repo is automatically updated from the source repo as part of the
release process.

To modify the formula for gitsh, submit a pull request against the gitsh
repo, specifcally `homebrew/gitsh.rb.in`.

For liftoff, the file is `homebrew/liftoff.rb` in the liftoff repo.

And in rcm modify `homebrew/rcm.rb.in` in the rcm repo.

Any pull requests against formulae in here will be closed.