Update complexity to 0.4.2 to allow for installation on M1 Macs
What?
=====

This introduces two changes necessary for installation on M1 Macs:

1. Bump complexity to 0.4.2
2. Introduce a compilation option, `--without-mimalloc`, which will
   instead use Rust's default allocator, which was made in the 0.4.2
   release of complexity
   (https://github.com/thoughtbot/complexity/releases/tag/0.4.2)

The underlying issue driving this change is MiMalloc does not work with
ARM architectures currently; this incorporates a change in complexity
allowing for MiMalloc to be an optional feature, which we carry into
this Homebrew formula to pass that configuration option along to folks
installing the tool that way.